Update Horizon Scanner.
the newer version of Horizon Scanner. more complete.

the outlier removal part has been added.
the attitude calculation method has been updated.
the uncertainty analysis has been included.
'roll' is the value of roll. 'uroll' is the uncertainty of roll.
'pitch' is the value of pitch. 'upitch' is the uncertainty of pitch.

one more update was needed to reduce the error in pitch calculation.